+++ templates/jp-mining-note/pa_sent/common.js
@@ -12,33 +12,29 @@
 
 /// {% set run %}
 
-//var sentences = document.querySelectorAll(".expression--sentence")
-//var isAltDisplay = false;
-//var isClozeDeletion = false;
+{
+  let isAltDisplay = false;
+  /* {% call IF('AltDisplayPASentenceCard') %} */
+    isAltDisplay = true;
+  /* {% endcall %} */
 
-//isAltDisplay = !!'{{ utils.any_of_str("AltDisplay") }}';
-
-var isAltDisplay = false;
-/* {% call IF('AltDisplayPASentenceCard') %} */
-  isAltDisplay = true;
-/* {% endcall %} */
-
-/* {% call IFNOT('AltDisplayPASentenceCard') %} */
-  /* {% call IF('AltDisplay') %} */
-  isAltDisplay = (
-    /* {% call utils.none_of('IsClickCard', 'IsHoverCard', 'IsSentenceCard', 'IsTargetedSentenceCard') %} */
-      false &&
+  /* {% call IFNOT('AltDisplayPASentenceCard') %} */
+    /* {% call IF('AltDisplay') %} */
+    isAltDisplay = (
+      /* {% call utils.none_of('IsClickCard', 'IsHoverCard', 'IsSentenceCard', 'IsTargetedSentenceCard') %} */
+        false &&
+      /* {% endcall %} */
+      true ? true : false
+    );
     /* {% endcall %} */
-    true ? true : false
-  );
+
+    /* {% call IFNOT('AltDisplay') %} */
+      isAltDisplay = false;
+    /* {% endcall %} */
   /* {% endcall %} */
 
-  /* {% call IFNOT('AltDisplay') %} */
-    isAltDisplay = false;
-  /* {% endcall %} */
-/* {% endcall %} */
-
-processSentences(isAltDisplay);
+  processSentences(isAltDisplay);
+}
 
 
 /// {% endset %}