+++ CHANGELOG.md
@@ -16,8 +16,14 @@ subsequent bumps are when the card has changed enough that a bump should be
 signified.)
 
 
+## [0.10.0.0] - 2022-08-??
 
-## [0.9.0.0] - 2022-08-??
+#### Changed (BREAKING)
+- Added `FrequencySort` field
+
+
+
+## [0.9.0.0] - 2022-08-28
 
 #### Changed (BREAKING)
 - yomichan templates: